- PostDetail screen is designed

+++ week5/src/main/java/com/chichi289/week5/data/remote/service/PostService.kt
@@ -25,7 +25,7 @@ interface PostService {
 
     @GET("v1/posts/{postId}")
     suspend fun getPostDetail(
-        @Path("postId") postId: Int
+        @Path("postId") postId: Long
     ): Response<PostDetailResponse>
 
     @DELETE("v1/posts/{postId}")
